Resolve review comments

+++ src/solver/conv_multipass_wino3x3WrW.cpp
@@ -496,7 +496,7 @@ bool ConvWinograd3x3MultipassWrW<WinoDataH, WinoFilterH, WinoDataW, WinoFilterW>
 
     // clang-format off
     {
-        const int64_t input_line_size = 4LL * problem.GetInWidth();
+        const int64_t input_line_size = static_cast<int64_t>(4) * problem.GetInWidth();
         const int64_t input_feature_map_size = input_line_size * problem.GetInHeight();
         const int64_t input_stack_size = input_feature_map_size * problem.GetInChannels();
         if (! (input_stack_size < (1U << 24)))
